+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed the error message when using the app on a non-zillow domain.
 - Aesthetic changes.
 - New selectors.
+- Friendly neighborhood google analytics up and running!!
 
 ## [Released]